+++ collector-amazon-ivs/src/test/java/com/bitmovin/analytics/amazon/ivs/manipulators/PlaybackEventDataManipulatorTest.kt
@@ -2,7 +2,6 @@ package com.bitmovin.analytics.amazon.ivs.manipulators
 
 import com.amazonaws.ivs.player.Player
 import com.bitmovin.analytics.amazon.ivs.TestUtils
-import com.bitmovin.analytics.data.MetadataProvider
 import io.mockk.every
 import io.mockk.mockk
 import org.assertj.core.api.Assertions.assertThat
@@ -11,12 +10,12 @@ import org.junit.Test
 internal class PlaybackEventDataManipulatorTest {
 
     @Test
-    fun `manipulate should set is live from sourceMetadata`() {
+    fun `manipulate should set is live from player`() {
         // arrange
-        val mockedMetadataProvider = mockk<MetadataProvider>(relaxed = true)
-        every { mockedMetadataProvider.getSourceMetadata()?.isLive } returns true
+        val mockedPlayer = mockk<Player>(relaxed = true)
+        every { mockedPlayer.duration } returns -1L
 
-        val manipulator = createPlaybackEventDataManipulator(metadataProvider = mockedMetadataProvider)
+        val manipulator = createPlaybackEventDataManipulator(mockedPlayer)
         val eventData = TestUtils.createMinimalEventData()
 
         // act
@@ -27,28 +26,24 @@ internal class PlaybackEventDataManipulatorTest {
     }
 
     @Test
-    fun `manipulate should set is live from player`() {
+    fun `manipulate should set is live from player to false`() {
         // arrange
         val mockedPlayer = mockk<Player>(relaxed = true)
-        every { mockedPlayer.duration } returns -1L
+        every { mockedPlayer.duration } returns 1234
 
-        val mockedMetadataProvider = mockk<MetadataProvider>(relaxed = true)
-        every { mockedMetadataProvider.getSourceMetadata()?.isLive } returns null
-
-        val manipulator = createPlaybackEventDataManipulator(mockedPlayer, mockedMetadataProvider)
+        val manipulator = createPlaybackEventDataManipulator(mockedPlayer)
         val eventData = TestUtils.createMinimalEventData()
 
         // act
         manipulator.manipulate(eventData)
 
         // assert
-        assertThat(eventData.isLive).isTrue
+        assertThat(eventData.isLive).isFalse
     }
 
     private fun createPlaybackEventDataManipulator(
         player: Player = mockk(relaxed = true),
-        metadataProvider: MetadataProvider = mockk(relaxed = true),
     ): PlaybackEventDataManipulator {
-        return PlaybackEventDataManipulator(player, metadataProvider)
+        return PlaybackEventDataManipulator(player)
     }
 }