+++ collector/src/main/java/com/bitmovin/analytics/data/MetadataProvider.kt
@@ -5,12 +5,12 @@ import com.bitmovin.analytics.api.DefaultMetadata
 import com.bitmovin.analytics.api.SourceMetadata
 import com.bitmovin.analytics.utils.ApiV3Utils
 import java.util.concurrent.ConcurrentHashMap
+import java.util.concurrent.atomic.AtomicReference
 
 class MetadataProvider {
     private val sourceMetadataMap = ConcurrentHashMap<Any, SourceMetadata?>()
-
-    // TODO: we should probably use thread save references for default metadata and deprecated bitmovin analytics config
-    var deprecatedBitmovinAnalyticsConfig: BitmovinAnalyticsConfig? = null
+    private val internalDefaultMetadata: AtomicReference<DefaultMetadata?> = AtomicReference(null)
+    private val deprecatedBitmovinAnalyticsConfig: AtomicReference<BitmovinAnalyticsConfig?> = AtomicReference(null)
 
     fun setSourceMetadata(source: Any, sourceMetadata: SourceMetadata?) {
         sourceMetadataMap[source] = sourceMetadata
@@ -20,15 +20,16 @@ class MetadataProvider {
     // in case there is no source metadata
     fun getSourceMetadata(source: Any): SourceMetadata? {
         val sourceMetadata = sourceMetadataMap[source]
-        val config = deprecatedBitmovinAnalyticsConfig
-
         if (sourceMetadata != null) {
             return sourceMetadata
-        } else if (config != null) {
-            return ApiV3Utils.extractSourceMetadata(config)
-        } else {
-            return null
         }
+
+        val config = deprecatedBitmovinAnalyticsConfig.get()
+        if (config != null) {
+            return ApiV3Utils.extractSourceMetadata(config)
+        }
+
+        return null
     }
 
     fun getSourceMetadata(): SourceMetadata? {
@@ -39,23 +40,29 @@ class MetadataProvider {
         setSourceMetadata(DEFAULT_KEY, sourceMetadata)
     }
 
-    // TODO: better naming, or check how i can use a backing field easier
-    private var internalDefaultMetadata: DefaultMetadata? = null
-
+    // For backwards compatibility reason we extract the data from the deprecatedAnalyticsConfig
+    // in case there is no defaultMetadata
     var defaultMetadata: DefaultMetadata
         get() {
-            val explicitSetDefaultMetadata = internalDefaultMetadata
-            return if (explicitSetDefaultMetadata != null) {
-                explicitSetDefaultMetadata
-            } else if (deprecatedBitmovinAnalyticsConfig != null) {
-                ApiV3Utils.extractDefaultMetadata(deprecatedBitmovinAnalyticsConfig!!)
-            } else {
-                DefaultMetadata()
+            val explicitSetDefaultMetadata = internalDefaultMetadata.get()
+            if (explicitSetDefaultMetadata != null) {
+                return explicitSetDefaultMetadata
             }
+
+            val config = deprecatedBitmovinAnalyticsConfig.get()
+            if (config != null) {
+                return ApiV3Utils.extractDefaultMetadata(config)
+            }
+
+            return DefaultMetadata()
         }
         set(value) {
-            internalDefaultMetadata = value
+            internalDefaultMetadata.set(value)
         }
+
+    fun setDeprectedBitmovinAnalyticsConfig(bitmovinAnalyticsConfig: BitmovinAnalyticsConfig) {
+        deprecatedBitmovinAnalyticsConfig.set(bitmovinAnalyticsConfig)
+    }
 
     companion object {
         private val DEFAULT_KEY = Any()